chat-assistant/backend/embedder-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
aiokafka==0.8.0
pinecone-client==5.2.1
httpx==0.24.1
python-dotenv==1.0.0
aiohttp==3.8.5
tenacity==8.2.2
sentence-transformers==2.7.0
torch>=2.0.0
numpy